+++ services.tf
@@ -1,27 +1,38 @@
 # All kubernetes network services
 
+
+locals {
+  _postgres_cluster_service = var.openwisp_services.setup_database ? [{
+    "port" : 5432, "name" : "postgres", "app" : "openwisp-postgres"
+  }] : []
+
+  _cluster_services = [
+    { "port" : 8000, "name" : "dashboard", "app" : "openwisp-dashboard" },
+    { "port" : 8001, "name" : "controller", "app" : "openwisp-controller" },
+    { "port" : 8002, "name" : "radius", "app" : "openwisp-radius" },
+    { "port" : 8003, "name" : "topology", "app" : "openwisp-topology" },
+    { "port" : 8004, "name" : "websocket", "app" : "openwisp-websocket" },
+    { "port" : 6379, "name" : "redis", "app" : "redis" },
+    { "port" : 25, "name" : "postfix", "app" : "openwisp-postfix" },
+    { "port" : 80, "name" : "dashboard-internal", "app" : "openwisp-nginx" },
+    { "port" : 80, "name" : "controller-internal", "app" : "openwisp-nginx" },
+    { "port" : 80, "name" : "radius-internal", "app" : "openwisp-nginx" },
+    { "port" : 80, "name" : "topology-internal", "app" : "openwisp-nginx" }
+  ]
+  cluster_services = concat(local._cluster_services, local._postgres_cluster_service)
+}
+
 resource "kubernetes_service" "openwisp_cluster_ip" {
   depends_on = [var.ow_cluster_ready]
-  for_each = {
-    0 : { "port" : 8000, "name" : "dashboard", "app" : "openwisp-dashboard" }
-    1 : { "port" : 8001, "name" : "controller", "app" : "openwisp-controller" }
-    2 : { "port" : 8002, "name" : "radius", "app" : "openwisp-radius" }
-    3 : { "port" : 8003, "name" : "topology", "app" : "openwisp-topology" }
-    4 : { "port" : 8004, "name" : "daphne", "app" : "openwisp-daphne" }
-    5 : { "port" : 5432, "name" : "postgres", "app" : "openwisp-postgres" }
-    6 : { "port" : 6379, "name" : "redis", "app" : "redis" }
-    7 : { "port" : 25, "name" : "postfix", "app" : "openwisp-postfix" }
-    8 : { "port" : 80, "name" : "dashboard-internal", "app" : "openwisp-nginx" }
-    9 : { "port" : 80, "name" : "controller-internal", "app" : "openwisp-nginx" }
-  }
+  count      = length(local.cluster_services)
   metadata {
-    name   = each.value.name
-    labels = { app = each.value.app }
+    name   = local.cluster_services[count.index].name
+    labels = { app = local.cluster_services[count.index].app }
   }
   spec {
     type     = "ClusterIP"
-    selector = { app = each.value.app }
-    port { port = each.value.port }
+    selector = { app = local.cluster_services[count.index].app }
+    port { port = local.cluster_services[count.index].port }
   }
 }
 
@@ -48,13 +59,13 @@ resource "kubernetes_service" "openwisp_nodeport" {
 locals {
   _use_openvpn = var.openwisp_services.use_openvpn ? [{
     "name" : "openvpn", "hostname" : "openvpn.openwisp.org",
-    "app" : "openwisp-openvpn", "load_balancer_ip" : var.infrastructure_provider.openvpn_loadbalancer_address,
+    "app" : "openwisp-openvpn", "load_balancer_ip" : var.infrastructure.openvpn_loadbalancer_address,
     "ports" : [
       { "name" : "vpn", "protocol" : "UDP", "port" : 1194 }
   ] }] : []
   _use_freeradius = var.openwisp_services.use_freeradius ? [{
     "name" : "freeradius", "hostname" : "freeradius.openwisp.org",
-    "app" : "openwisp-freeradius", "load_balancer_ip" : var.infrastructure_provider.freeradius_loadbalancer_address,
+    "app" : "openwisp-freeradius", "load_balancer_ip" : var.infrastructure.freeradius_loadbalancer_address,
     "ports" : [
       { "name" : "auth", "protocol" : "UDP", "port" : 1812 },
       { "name" : "acct", "protocol" : "UDP", "port" : 1813 }
@@ -93,7 +104,7 @@ resource "kubernetes_ingress" "http_ingress" {
 
   metadata {
     name        = "openwisp-http-ingress"
-    annotations = { "kubernetes.io/ingress.global-static-ip-name" = var.infrastructure_provider.http_loadbalancer_name }
+    annotations = { "kubernetes.io/ingress.global-static-ip-name" = var.infrastructure.http_loadbalancer_name }
     labels      = { "app" = "openwisp-nginx" }
   }
   spec {