+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- openwisp-websocket container
+- pinned openwisp images to 'alpha.1'
+- support for external database
+- support for different sslmodes for database connections
 
 ## [0.1.0-alpha.2] - 2020-02-12
 ### Added